pyproject.toml
[project]
name = "solarphunk"
version = "0.1.0"
description = "Autonomous solarpunk blog — research, dream, publish"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.30.0",
    "httpx>=0.27.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pyyaml>=6.0.0",
    "python-slugify>=8.0.0",
    "pydantic>=2.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "lib"]
